geolocation/getCurrentPosition_TypeError.https.html

<!DOCTYPE html>
<meta charset="utf-8" />
<title>Geolocation Test: getCurrentPosition TypeError tests</title>
<link rel="help" href="http://www.w3.org/TR/geolocation-API/" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
  test(() => {
    assert_throws_js(TypeError, () => {
      navigator.geolocation.getCurrentPosition();
    });
  }, "Call getCurrentPosition without arguments, check that exception is thrown");

  test(() => {
    assert_throws_js(TypeError, () => {
      navigator.geolocation.getCurrentPosition(null);
    });
  }, "Call getCurrentPosition with null success callback, check that exception is thrown");

  test(() => {
    assert_throws_js(TypeError, () => {
      navigator.geolocation.getCurrentPosition(null, null);
    });
  }, "Call getCurrentPosition with null success and error callbacks, check that exception is thrown");

  test(() => {
    assert_throws_js(TypeError, () => {
      navigator.geolocation.getCurrentPosition(3);
    });
  }, "Call getCurrentPosition() with wrong type for first argument. Exception expected.");

  test(() => {
    assert_throws_js(TypeError, () => {
      navigator.geolocation.getCurrentPosition(() => {}, 4);
    });
  }, "Call getCurrentPosition() with wrong type for second argument. Exception expected.");

  test(() => {
    assert_throws_js(TypeError, () => {
      navigator.geolocation.getCurrentPosition(
        () => {},
        () => {},
        4
      );
    });
  }, "Call getCurrentPosition() with wrong type for third argument. Exception expected.");

  test(() => {
    const handleEvent = () => {};
    assert_throws_js(TypeError, () => {
      navigator.geolocation.getCurrentPosition({ handleEvent });
    });

    assert_throws_js(TypeError, () => {
      navigator.geolocation.getCurrentPosition(() => {}, { handleEvent });
    });
  }, "Calling getCurrentPosition() with a legacy event handler objects.");
</script>
